separate data aggregator from xml printer.
